chore: update type defs
